// agent/src/common/generated/proto/key_pool.ts
export enum HashFunction {
  /** HASH_FUNCTION_UNSPECIFIED - Reserve 0 to detect unset values. */
  HASH_FUNCTION_UNSPECIFIED = 0,
  /** HASH_FUNCTION_SHA256 - Commonly used (e.g. Bitcoin). */
  HASH_FUNCTION_SHA256 = 1,
  /** HASH_FUNCTION_KECCAK256 - Used by Ethereum (original Keccak, not SHA3). */
  HASH_FUNCTION_KECCAK256 = 2,
  /** HASH_FUNCTION_SHA3_256 - NIST standardized versions of Keccak. */
  HASH_FUNCTION_SHA3_256 = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface SigningKey {
  /**
   * / The semantics of some values are specified in `BuiltinSigningKey` -
   * / in particular, zero has the special meaning of unspecified/default.
   * /
   * / Values outside the range defined in this enum are application specific.
   * /
   * / The maximum allowed value depends on the configuration. Note that
   * / if the server is configured to support N secret keys, these are
   * / accessed using key_index 1..N.
   */
  keyIndex: number;
}
// ...
export interface SignMessageRequest {
  /** / Default key: `SIGNING_KEY_SERVICE_RESPONSE` (used if left as `SIGNING_KEY_UNSPECIFIED`). */
  signingKey:
    | SigningKey
    | undefined;
  /** / Hash function to use to compute the digest to sign. */
  hashFunction: HashFunction;
  /** / The bytes of the message to sign. Maximum message size is 1Mib (2**20). */
  message: Buffer;
}

export interface SignMessageResponse {
  /**
   * / An signatuer using Ethereum's 65 byte format `r || s || v`, where `r` and `s`
   * / the signature components (each 32 bytes) and `v` is a zero or one byte
   * / used for key recovery.
   */
  signature: Buffer;
}

export interface SignEthereumTransactionRequest {
  /** / Default key: `SIGNING_KEY_ETHEREUM` (used if left as `SIGNING_KEY_UNSPECIFIED`). */
  signingKey:
    | SigningKey
    | undefined;
  /**
   * / RLP-encoded unsigned transaction.
   * / Must contain 6 or 9 element. If 9 elements are used, the 7th element
   * / it is assumed to be the chain ID and the transaction is signed using EIP-155;
   * / otherwise a legacy signature is used.
   */
  txData: Buffer;
}

export interface SignEthereumTransactionResponse {
  /**
   * / RLP-encoded signed transaction.
   * / Always 9 elements: the six elements from the input, followed by v, r, s.
   */
  txData: Buffer;
}

export interface GetEthereumAddressRequest {
  /** / Default key: `SIGNING_KEY_ETHEREUM` (used if left as `SIGNING_KEY_UNSPECIFIED`). */
  signingKey: SigningKey | undefined;
}
// ...
export const SigningKey: MessageFns<SigningKey> = {
  encode(message: SigningKey, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.keyIndex !== 0) {
      writer.uint32(8).uint32(message.keyIndex);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SigningKey {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSigningKey();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.keyIndex = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SigningKey {
    return { keyIndex: isSet(object.keyIndex) ? globalThis.Number(object.keyIndex) : 0 };
  },

  toJSON(message: SigningKey): unknown {
    const obj: any = {};
    if (message.keyIndex !== 0) {
      obj.keyIndex = Math.round(message.keyIndex);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SigningKey>, I>>(base?: I): SigningKey {
    return SigningKey.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SigningKey>, I>>(object: I): SigningKey {
    const message = createBaseSigningKey();
    message.keyIndex = object.keyIndex ?? 0;
    return message;
  },
};
// ...
export const SignMessageRequest: MessageFns<SignMessageRequest> = {
  encode(message: SignMessageRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signingKey !== undefined) {
      SigningKey.encode(message.signingKey, writer.uint32(10).fork()).join();
    }
    if (message.hashFunction !== 0) {
      writer.uint32(16).int32(message.hashFunction);
    }
    if (message.message.length !== 0) {
      writer.uint32(26).bytes(message.message);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SignMessageRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSignMessageRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signingKey = SigningKey.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.hashFunction = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.message = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SignMessageRequest {
    return {
      signingKey: isSet(object.signingKey) ? SigningKey.fromJSON(object.signingKey) : undefined,
      hashFunction: isSet(object.hashFunction) ? hashFunctionFromJSON(object.hashFunction) : 0,
      message: isSet(object.message) ? Buffer.from(bytesFromBase64(object.message)) : Buffer.alloc(0),
    };
  },

  toJSON(message: SignMessageRequest): unknown {
    const obj: any = {};
    if (message.signingKey !== undefined) {
      obj.signingKey = SigningKey.toJSON(message.signingKey);
    }
    if (message.hashFunction !== 0) {
      obj.hashFunction = hashFunctionToJSON(message.hashFunction);
    }
    if (message.message.length !== 0) {
      obj.message = base64FromBytes(message.message);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SignMessageRequest>, I>>(base?: I): SignMessageRequest {
    return SignMessageRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SignMessageRequest>, I>>(object: I): SignMessageRequest {
    const message = createBaseSignMessageRequest();
    message.signingKey = (object.signingKey !== undefined && object.signingKey !== null)
      ? SigningKey.fromPartial(object.signingKey)
      : undefined;
    message.hashFunction = object.hashFunction ?? 0;
    message.message = object.message ?? Buffer.alloc(0);
    return message;
  },
};
// ...
export const SignEthereumTransactionRequest: MessageFns<SignEthereumTransactionRequest> = {
  encode(message: SignEthereumTransactionRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signingKey !== undefined) {
      SigningKey.encode(message.signingKey, writer.uint32(10).fork()).join();
    }
    if (message.txData.length !== 0) {
      writer.uint32(18).bytes(message.txData);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SignEthereumTransactionRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSignEthereumTransactionRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signingKey = SigningKey.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.txData = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SignEthereumTransactionRequest {
    return {
      signingKey: isSet(object.signingKey) ? SigningKey.fromJSON(object.signingKey) : undefined,
      txData: isSet(object.txData) ? Buffer.from(bytesFromBase64(object.txData)) : Buffer.alloc(0),
    };
  },

  toJSON(message: SignEthereumTransactionRequest): unknown {
    const obj: any = {};
    if (message.signingKey !== undefined) {
      obj.signingKey = SigningKey.toJSON(message.signingKey);
    }
    if (message.txData.length !== 0) {
      obj.txData = base64FromBytes(message.txData);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SignEthereumTransactionRequest>, I>>(base?: I): SignEthereumTransactionRequest {
    return SignEthereumTransactionRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SignEthereumTransactionRequest>, I>>(
    object: I,
  ): SignEthereumTransactionRequest {
    const message = createBaseSignEthereumTransactionRequest();
    message.signingKey = (object.signingKey !== undefined && object.signingKey !== null)
      ? SigningKey.fromPartial(object.signingKey)
      : undefined;
    message.txData = object.txData ?? Buffer.alloc(0);
    return message;
  },
};
// ...
export const GetEthereumAddressRequest: MessageFns<GetEthereumAddressRequest> = {
  encode(message: GetEthereumAddressRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signingKey !== undefined) {
      SigningKey.encode(message.signingKey, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetEthereumAddressRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetEthereumAddressRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signingKey = SigningKey.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetEthereumAddressRequest {
    return { signingKey: isSet(object.signingKey) ? SigningKey.fromJSON(object.signingKey) : undefined };
  },

  toJSON(message: GetEthereumAddressRequest): unknown {
    const obj: any = {};
    if (message.signingKey !== undefined) {
      obj.signingKey = SigningKey.toJSON(message.signingKey);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetEthereumAddressRequest>, I>>(base?: I): GetEthereumAddressRequest {
    return GetEthereumAddressRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetEthereumAddressRequest>, I>>(object: I): GetEthereumAddressRequest {
    const message = createBaseGetEthereumAddressRequest();
    message.signingKey = (object.signingKey !== undefined && object.signingKey !== null)
      ? SigningKey.fromPartial(object.signingKey)
      : undefined;
    return message;
  },
};
// ...
export type KeyPoolServiceService = typeof KeyPoolServiceService;
// ...
export const KeyPoolServiceClient = makeGenericClientConstructor(
  KeyPoolServiceService,
  "key_pool.KeyPoolService",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): KeyPoolServiceClient;
  service: typeof KeyPoolServiceService;
  serviceName: string;
};

function bytesFromBase64(b64: string): Uint8Array {
  return Uint8Array.from(globalThis.Buffer.from(b64, "base64"));
}

function base64FromBytes(arr: Uint8Array): string {
  return globalThis.Buffer.from(arr).toString("base64");
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}